Adds getPlayerChoice function
This adds the player choice function to the
javaScript file. This uses the prompt to let the
user enter in a string of either rock paper or
scissors. The funciton then validates the input
as only those three options and returns that string
with the first letter capatalized and the rest of
them lowerCase.